+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/document_settings.rb
@@ -7,7 +7,9 @@ module OoxmlParser
     # Parse Settings object
     # @return [DocumentSettings] result of parsing
     def parse
-      doc = Nokogiri::XML(File.open(OOXMLDocumentObject.path_to_folder + 'word/settings.xml'))
+      settings_path = OOXMLDocumentObject.path_to_folder + 'word/settings.xml'
+      return nil unless File.exist?(settings_path)
+      doc = Nokogiri::XML(File.open(settings_path))
       doc.xpath('w:settings/*').each do |node_child|
         case node_child.name
         when 'defaultTabStop'