+++ spec/document/settings_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+describe 'DocumentSettings' do
+  it 'no_settings_xml' do
+    docx = OoxmlParser::Parser.parse('spec/document/settings/no_setting_xml.docx')
+    expect(docx.settings).to be_nil
+  end
+end