+++ Changelog.md
@@ -18,6 +18,7 @@
   
 ### Fixes
 * Fix crash on empty coordinates list of chart
+* Fix crash on docx with no `settings.xml`
 
 ## 0.4.1 (2018-03-01)
 ### Fixes